backend/requirements.txt
# Core dependencies for simple GPT-4o bill analyzer
openai==1.35.0
Pillow==10.1.0
python-dotenv==1.0.0

# Optional: If you want to keep FastAPI
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6
